move zero_utils.hpp to the utils folder
