+++ src/plugins/intel_npu/src/backend/include/zero_infer_request.hpp
@@ -14,7 +14,7 @@
 #include "zero_pipeline.hpp"
 #include "zero_profiling.hpp"
 #include "zero_remote_tensor.hpp"
-#include "zero_utils.hpp"
+#include "intel_npu/utils/zero/zero_utils.hpp"
 #include "zero_wrappers.hpp"
 
 namespace intel_npu {
