+++ src/plugins/intel_npu/src/backend/src/zero_init.cpp
@@ -6,9 +6,9 @@
 
 #include "intel_npu/common/itt.hpp"
 #include "intel_npu/utils/zero/zero_api.hpp"
+#include "intel_npu/utils/zero/zero_utils.hpp"
 #include "ze_api.h"
 #include "ze_command_queue_npu_ext.h"
-#include "zero_utils.hpp"
 
 #ifdef _WIN32
 namespace {
