+++ src/plugins/intel_npu/src/backend/src/zero_remote_tensor.cpp
@@ -8,8 +8,8 @@
 
 #include "intel_npu/config/common.hpp"
 #include "intel_npu/utils/zero/zero_api.hpp"
+#include "intel_npu/utils/zero/zero_utils.hpp"
 #include "openvino/core/type/element_iterator.hpp"
-#include "zero_utils.hpp"
 
 using namespace ov::intel_npu;
