+++ src/plugins/intel_npu/src/utils/include/intel_npu/utils/zero/zero_utils.hpp
